pyproject.toml
[project]
name = "labloom"
version = "0.1.0"
description = "AI lab generation tool for Virtual Labs"
readme = "README.md"
authors = [
    { name = "Nitheesh Chandra", email = "nitheeshchandra.y@research.iiit.ac.in" }
]
classifiers = ["Private :: Do Not Upload"]
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.0.4",
    "google-genai>=1.20.0",
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.8",
    "langgraph-checkpoint-sqlite>=2.0.10",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "flake8>=7.2.0",
    "mypy>=1.16.0",
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.11.13",
    "tox>=4.26.0",
]

[project.scripts]
labloom = "labloom.__main__:cli"

[project.gui-scripts]
labloom-ui = "labloom.__main__:gui"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"



[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=labloom",
    "--cov-report=term-missing",
    # "--cov-report=html:coverage_html_report",
    # "--cov-report=xml:coverage.xml",
]

[tool.flake8]
max-line-length = 128

[tool.mypy]
mypy_path = "src"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
no_implicit_reexport = true
